+++ docs/changes.md
@@ -5,6 +5,11 @@ title: Memory Change Log
 
 ## Version 3.2 {#v3-2}
 
+### v3.2.2 {#v3-2-2}
+
+* Add support for new `Orchestra\Memory\Provider::secureGet()` and `Orchestra\Memory\Provider::secureSet()`.
+* Add `Orchestra\Memory\Provider::securePut()` as an encrypted alias to `Orchestra\Memory\Provider::put()`.
+
 ### v3.2.1 {#v3-2-1}
 
 * Ignores `PDOException` exception when saving database based driver as the code maybe executed from outside of configured environment.
@@ -15,6 +20,11 @@ title: Memory Change Log
 * Improved performances by reducing call within `Illuminate\Container\Container`.
 
 ## Version 3.1 {#v3-1}
+
+### v3.1.3 {#v3-1-3}
+
+* Add support for new `Orchestra\Memory\Provider::secureGet()` and `Orchestra\Memory\Provider::secureSet()`.
+* Add `Orchestra\Memory\Provider::securePut()` as an encrypted alias to `Orchestra\Memory\Provider::put()`.
 
 ### v3.1.2 {#v3-1-2}
 
@@ -29,6 +39,11 @@ title: Memory Change Log
 * Update support to Laravel Framework v5.1.
 
 ## Version 3.0 {#v3-0}
+
+### v3.0.6 {#v3-0-6}
+
+* Add support for new `Orchestra\Memory\Provider::secureGet()` and `Orchestra\Memory\Provider::secureSet()`.
+* Add `Orchestra\Memory\Provider::securePut()` as an encrypted alias to `Orchestra\Memory\Provider::put()`.
 
 ### v3.0.5 {#v3-0-5}